+++ pcs/lib/pacemaker/live.py
@@ -3,6 +3,7 @@ import re
 from typing import (
     Iterable,
     List,
+    Optional,
     Tuple,
 )
 
@@ -417,7 +418,12 @@ def remove_node(runner, node_name):
 ### resources
 
 def resource_cleanup(
-    runner, resource=None, node=None, operation=None, interval=None
+    runner: CommandRunner,
+    resource: Optional[str] = None,
+    node: Optional[str] = None,
+    operation: Optional[str] = None,
+    interval: Optional[str] = None,
+    strict: bool = False,
 ):
     cmd = [__exec("crm_resource"), "--cleanup"]
     if resource:
@@ -428,6 +434,8 @@ def resource_cleanup(
         cmd.extend(["--operation", operation])
     if interval:
         cmd.extend(["--interval", interval])
+    if strict:
+        cmd.extend(["--force"])
 
     stdout, stderr, retval = runner.run(cmd)
 
@@ -442,7 +450,13 @@ def resource_cleanup(
     # usefull output (what has been done) goes to stderr
     return join_multilines([stdout, stderr])
 
-def resource_refresh(runner, resource=None, node=None, full=False, force=None):
+def resource_refresh(
+    runner: CommandRunner,
+    resource: Optional[str] = None,
+    node: Optional[str] = None,
+    strict: bool = False,
+    force: bool = False,
+):
     if not force and not node and not resource:
         summary = ClusterState(get_cluster_status_xml(runner)).summary
         operations = summary.nodes.attrs.count * summary.resources.attrs.count
@@ -458,7 +472,7 @@ def resource_refresh(runner, resource=None, node=None, full=False, force=None):
         cmd.extend(["--resource", resource])
     if node:
         cmd.extend(["--node", node])
-    if full:
+    if strict:
         cmd.extend(["--force"])
 
     stdout, stderr, retval = runner.run(cmd)