+++ pcs_test/tools/command_env/config_runner_pcmk.py
@@ -461,6 +461,7 @@ class PcmkShortcuts():
         before=None,
         resource=None,
         node=None,
+        strict=False,
         stdout="",
         stderr="",
         returncode=0
@@ -475,6 +476,7 @@ class PcmkShortcuts():
             placed
         string resource -- the id of a resource to be cleaned
         string node -- the name of the node where resources should be cleaned
+        bool strict -- strict mode of 'crm_resource cleanup' enabled?
         string stdout -- crm_resource's stdout
         string stderr -- crm_resource's stderr
         int returncode -- crm_resource's returncode
@@ -484,6 +486,8 @@ class PcmkShortcuts():
             cmd.extend(["--resource", resource])
         if node:
             cmd.extend(["--node", node])
+        if strict:
+            cmd.extend(["--force"])
         self.__calls.place(
             name,
             RunnerCall(